requirements.txt
streamlit
selenium
beautifulsoup4
python-dotenv
langchain-core
langchain-mistralai
langgraph
typing-extensions